rpmsg: qcom: glink: support waking up on channel rx
Configure all channels as wakeup capable and report a wakeup event
when data is received.

This allows userspace to "subscribe" to a particular channel where
it is useful to wake up to process new data. The expected usecase
is to allow for handling incoming SMS or phone calls where the only
notification mechanism is via QRTR on the IPCRTR glink channel.

As this behaviour is likely undesirable for most users, this patch
doesn't enable a wakeup_source for any channels by default.

Signed-off-by: Casey Connolly <casey.connolly@linaro.org>

FIX: rpmsg: glink: fix wakeup nullptr

Somehow rpdev can be null???

weird...